[CDH4] Add catalog APIs for adding/removing roles and privileges
Adds these APIs to the CatalogServiceCatalog so we can be sure that incrementing
the catalog version and the update happen atomically. Also moves the logic of
performing the incremental updates from Sentry out of the catalog and into a
new SentryServiceUpdater class.

CDH4 does not support Sentry Policy Service so all functionality is disabled.

Change-Id: I9c16cc48d1d10a0066bd04156db1433ba5d5d0e5
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4165
Reviewed-by: Lenni Kuff <lskuff@cloudera.com>
Tested-by: jenkins