Disable analytic fn query tests for avro
Avro doesn't support a timestamp type and the analytic fn tests
need to exercise avg() for timestamps because there's a template
specialization of the avg() UDA for timestamps. It's less
important right now to test on avro because the analytic function
evaluation is unlikely to be affected by file formats. There is
a TODO to enable some of the tests for avro later, if possible.

Change-Id: I5d69816cc0864deab4cf01b6a5969d39cec8875a
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4238
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Reviewed-by: Matthew Jacobs <mj@cloudera.com>
Tested-by: Matthew Jacobs <mj@cloudera.com>