PHJ: Adding probe-side filters.
Also, avoiding to build slot filters in case of joins on tables whose majority is not parquet.
Previously, we would build the slot filters in the build side even if the scanners on the
probe side would not use them (i.e. were not HdfsParquetScanners).

Note that currently in order to decide whether we are going to add probe-side filters or not,
we use the same heuristic with the old (HashJoinNode) code. But we can do better
because we now know the total number of rows and the size of each partition,
before we start building the hash tables.

Also note that this will work only if the first partitioning level of PHJ is CRC, the same
with the one the parquet scanners use.

Change-Id: Id290251db7daffe2c85943829b12b167ff0f1889
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4299
Reviewed-by: Ippokratis Pandis <ipandis@cloudera.com>
Tested-by: jenkins