IMPALA-1072: Parquet writer is misestimating the file size.
The file size estimate was double counting plain encoded values (and one of
the counts was the uncompressed size).  Now, the file size estimate is
computed as the running uncompressed size of the dictionary, along with the
size of all finalized data pages and their headers (after compression).  (The
file size limiting heuristic already accounts for some slop needed for the not
yet finalized pages).

So that the file sizes don't change too much as a result of the change,
lower the HDFS block size / max parquet file size to 512 GB.

Thanks to Nong for the test.

I had also tested this manually with the default block size to see that
parquet files written by impala are now just under 512 GB.

Change-Id: I6b09c528a7e2372d86bbb7692eea4624125cae14
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4347
Reviewed-by: Daniel Hecht <dhecht@cloudera.com>
Tested-by: jenkins