IMPALA-1151: Fix use-after-free in webserver
It's an error to add a temporary string directly to a RapidJson document
without wrapping it in a Value object first, because the latter will
copy the data whereas the former only moves it. Unfortunately, RapidJson
doesn't do anything to prevent this usage (which is, at least, safe for
literals).

Webserver::RootHandler() had several of these errors. I don't think we
ever saw a related crash, but if someone had hammered
http://<impalad>:25000/ there may have been an issue.

Change-Id: Ic3d5f6bdd7e60f4660f4eb713385664e76fb565f
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4360
Reviewed-by: Daniel Hecht <dhecht@cloudera.com>
Tested-by: jenkins