[CDH4] Add support for SHOW GRANT ROLE <roleName> [ON <privilege spec>]
This is a cherry-pick from CDH5. All GRANT/REVOKE statements have been disabled
because Sentry Service is not supported on CDH4.

Adds support for displaying all or a subset of the privileges granted to a role. Users
have privileges to execute this statement if they are already granted the role or if
they are an admin user on the Sentry Policy Service. The output includes:
* The target scope of the privilege
* The privilege level
* The target names in the object hierarchy
* Whether the privilege was granted using WITH GRANT OPTION
* The create time of the privilege

Examples:
-- Show all grants in role1
SHOW GRANT ROLE role1

-- Shows all grants in role1 on the database foo
SHOW GRANT ROLE role1 on DATABASE foo

Output looks like:
+----------+------------+-------+-----+-----------+--------------+-------------------------------+
| scope    | database   | table | uri | privilege | grant_option | create_time                   |
+----------+------------+-------+-----+-----------+--------------+-------------------------------+
| DATABASE | functional |       |     | ALL       | false        | Fri, Sep 19 2014 16:13:40.999 |
+----------+------------+-------+-----+-----------+--------------+-------------------------------+

Change-Id: I8ef1b87a4c22c8fba4228012668033d7f9d06fcb
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4389
Reviewed-by: Lenni Kuff <lskuff@cloudera.com>
Tested-by: jenkins