Ignore invalid query options when opening an HS2 session
The CDH5.2.0 JDBC driver tries to set a database with a query
option. Doing so would fail the whole session open call which returned
an error if there was any problem with the query options. This patch
changes that to be more forgiving - incorrect query options would still
be logged.

Despite being found on CDH5, this is good for CDH4 as well.

Change-Id: Iaeef06238cc7994af00cdcfe7401528cdc242a30
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4412
Tested-by: jenkins
Reviewed-by: Henry Robinson <henry@cloudera.com>