Fix bug when catalog object created and dropped within a statestore heartbeat using SYNC_DDL
If an object is created and dropped within a statestore heartbeat the statestore
will not send any update out mentioning that this item ever existed. If the object
was created with SYNC_DDL=0 and then dropped with SYNC_DDL=1, this is problematic because
the creation will be applied to the impalad catalog using the "direct-DDL" path but
the drop will not ever be applied because the statestore heartbeat we are waiting on will
not contain any drop message.
The fix is to always apply updates using the "direct-DDL" path, even if SYNC_DDL is
enabled. After applying the update, we will wait until all other nodes have seen the
update before continuing.

Change-Id: I0e04c81f50ba253d5aa97e8bee3568bae4ffdff1
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4421
Reviewed-by: Lenni Kuff <lskuff@cloudera.com>
Tested-by: jenkins