Don't check vcore subscription levels if scanner thread is reserved
Each scan node needs at least one thread to make progress. In a query
with a lot of scan nodes, the number of threads required becomes very
high, and for certain exec nodes (e.g. joins and unions) it's not ok to
run scan nodes serially, since they become blocked on the input from a
single scan that may not have yet run, and can't consume the input from
nodes that currently have an active scan thread.

The ThreadResourceMgr handles this case by allowing reserved thread
tokens that are always granted. However, with Llama enabled, we would
always check the thread-to-vcore ratio before allowing a scanner thread
to start without taking into account whether the thread was reserved or
not. This could lead to starvation, and ultimately a deadlocked query.

This patch allows threads to tell whether or not a token has come from a
reservation or not. The scanner threads can then use this information to
decide whether to check the subscription ratio before running the
thread.

Of course, if the subscription ratio really is that high, there'll be
other performance problems but at least the query should make (slow) progress.

Change-Id: I1ff0c94d919ce3a2b29380741a62f7b8368c6c59
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4614
Reviewed-by: Henry Robinson <henry@cloudera.com>
Tested-by: Henry Robinson <henry@cloudera.com>