IMPALA-1324,IMPALA-1307: Enforce compatible hash exprs for partitioned joins.
When generating the plan for a partitioned hash join we place the join into a
partition-compatible input fragment, if possible. During this optimization one
must ensure that the hash exchange sending to the new join (which was placed
into a compatible fragment) is compatible with the hash exprs used for sending
to the fragment containing the join node.
In particular, we had two bugs:
1. The number of hash exprs could be different, possibly because of redundant
   exprs on one or both sides
2. The order of the hash exprs could be different, causing two rows with the
   same hash-expr values to be sent to different nodes

The fix is to enfore the above two properties, and revert to exchanging
both sides if no compatible hash exprs can be constructed.

Change-Id: Id155fb8094ed1694f7bc038ed2f9685f4d645fbe
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4639
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: jenkins