CDH-22027: Remove default Llama hosts
There are two flags that allow us to specify Llama hosts - one was added
for HA, while the other is in the process of being deprecated. Both had
default values which interfered with each other if they weren't
overwritten. This patch removes the defaults, and Impala won't start if
neither are set.

Change-Id: Ibba771ab91ee13e5b5d1a725e385c964f2f9a132
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4652
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: jenkins